+++ tests/base_tests.py
@@ -7,7 +7,7 @@ from yaml.scanner import ScannerError
 
 from flask import Flask as BaseFlask
 from ordbok import Ordbok
-from ordbok.private import PrivateConfigFile
+from ordbok.config_private import PrivateConfigFile
 from ordbok.flask_helper import (
     Flask as OrdbokFlask, OrdbokFlaskConfig, make_config)
